backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
llama-index = "0.10.44"
llama-index-core = "0.10.44"
llama-index-agent-openai = "0.2.7"
openinference-semantic-conventions = "*"
openinference-instrumentation-llama-index = "^2.0.0"
opentelemetry-exporter-otlp = "^1.25.0"
requests = "^2.32.3"
torch = "^2.4.1"
torchvision = "^0.19.1"
transformers = "^4.45.2"
pillow = "^10.4.0"
python-multipart = "^0.0.12"
accelerate = ">=0.26.0"
huggingface-hub = "^0.25.2"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
